lambda_generated/requirements.txt
praw==7.5.0
requests==2.27.1
python-dotenv==0.19.2
